Started code refactoring
