+++ app/src/main/java/com/geocipher/app/implementations/MessageRepository.kt
@@ -1,0 +1,9 @@
+package com.geocipher.app.implementations
+
+import com.geocipher.app.interfaces.IMessageRepository
+
+class MessageRepository : IMessageRepository {
+    override fun addMessage(message: String, key: String, latitude: Double, longitude: Double) {
+        TODO("Not yet implemented")
+    }
+}
