+++ app/src/main/java/com/geocipher/app/interfaces/ILocationManager.kt
@@ -1,0 +1,7 @@
+package com.geocipher.app.interfaces
+
+interface ILocationManager {
+    fun getUserLocation()
+    fun requestLocationPermission()
+    fun checkLocationPermission(): Boolean
+}
