+++ app/src/main/java/com/geocipher/app/interfaces/IMessageRepository.kt
@@ -1,0 +1,5 @@
+package com.geocipher.app.interfaces
+
+interface IMessageRepository {
+    fun addMessage(message: String, key: String, latitude: Double, longitude: Double)
+}
