+++ app/src/main/java/com/geocipher/app/interfaces/LocationUpdateListener.kt
@@ -1,0 +1,6 @@
+package com.geocipher.app.interfaces
+
+interface LocationUpdateListener {
+    fun onLocationUpdated(latitude: Double, longitude: Double)
+    fun onLocationError(error: String)
+}
